API

Types / Constructors

PsychometricTests.PsychometricTest — Type

PsychometricTest{Ti,Tp,Tr,Ts}

A struct representing a psychometric test.

Missing docstring.

Missing docstring for PsychometricTest(::AbstractVector{<:Item}, ::AbstractVector{<:Person}, ::AbstractVector{<:Response}). Check Documenter's build log for details.

PsychometricTests.PsychometricTest — Method

PsychometricTest(m::AbstractMatrix; scales = nothing)

Construct a PsychometricTest from a response matrix m.

julia> m = ones(Int, 10, 2);

julia> PsychometricTest(m);

Missing docstring.

Missing docstring for PsychometricTest(::Any). Check Documenter's build log for details.

PsychometricTests.Item — Type

Item

An abstract type representing an item in a psychometric test.

Every implementation T <: Item must define the following interface:

getid: Get the unique item identifier.

PsychometricTests.Person — Type

Person

An abstract type representing a person in a psychometric test.

Every implementation T <: Person must define the following interface:

getid: Get the unique person identifier.

PsychometricTests.Response — Type

Response

An abstract type representing a response in a psychometric test.

Every implementation of T <: Response must define the following interface:

getvalue: Get the response value.

PsychometricTests.BasicItem — Type

BasicItem{I} <: Item

A minimal implementation of Item. Contains no information besides a unique id of type I.

PsychometricTests.BasicPerson — Type

BasicPerson{I} <: Person

A minimal implementation of Person. Contains no information besides a unique id of type I.

PsychometricTests.BasicResponse — Type

BasicResponse{T}

A minimal implementation of Response. Contains a response value value::T.

Missing docstring.

Missing docstring for Matrix. Check Documenter's build log for details.

Accessors

Missing docstring.

Missing docstring for getitems. Check Documenter's build log for details.

Missing docstring.

Missing docstring for getpersons. Check Documenter's build log for details.

Missing docstring.

Missing docstring for getresponses. Check Documenter's build log for details.

PsychometricTests.getvalue — Function

getvalue(response::Response)

Get the response value of response.

PsychometricTests.getid — Function

getid(item::Item)

Get the unique item identifier of item.

getid(person::Person)

Get the unique person identifier of person.

Missing docstring.

Missing docstring for getitemid. Check Documenter's build log for details.

Missing docstring.

Missing docstring for getpersonid. Check Documenter's build log for details.

Missing docstring.

Missing docstring for nitems. Check Documenter's build log for details.

Missing docstring.

Missing docstring for npersons. Check Documenter's build log for details.

Missing docstring.

Missing docstring for nresponses. Check Documenter's build log for details.

Iterators

Missing docstring.

Missing docstring for eachitem. Check Documenter's build log for details.

Missing docstring.

Missing docstring for eachperson. Check Documenter's build log for details.

Missing docstring.

Missing docstring for eachresponse. Check Documenter's build log for details.

Descriptive Statistics

PsychometricTests.itemmean — Function

itemmean(test::PsychometricTest, id)

Calculate the mean score for item with identifier id in test.

PsychometricTests.itemmeans — Function

itemmeans(test::PsychometricTest)
itemmeans(test::PsychometricTest, scale::Symbol)

Calculate the mean scores for each item in test.

PsychometricTests.itemscore — Function

itemscore(test::PsychometricTest, id)

Calculate the total score for item with identifier id in test.

PsychometricTests.itemscores — Function

itemscores(test::PsychometricTest)
itemscores(test::PsychometricTest, scale::Symbol)

Calculate the total score for each item in test.

PsychometricTests.personmean — Function

personmean(test::PsychometricTest, id)

Calculate the mean score for person with identifier id in test.

PsychometricTests.personmeans — Function

personmeans(test::PsychometricTest)

Calculate the mean scores for each person in test.

PsychometricTests.personscore — Function

personscore(test::PsychometricTest, id)

Calculate the total score for person with identifier id in test.

PsychometricTests.personscores — Function

personscores(test::PsychometricTest)
personscores(test::PsychometricTest, scale::Symbol)

Calculate the total scores for each person in test.