API

Types / Constructors

PsychometricTests.PsychometricTest — Type

PsychometricTest{I<:Item,P<:Person,R<:Response,IIT,PIT,ZT}

A struct representing a psychometric test.

Fields

items: A vector of unique items.

persons: A vector of unique persons.

responses: A vector of responsese.

item_ptr: A dictionary of key-value-pairs mapping the unique item identifier to responses.

person_ptr: A dictionary of key-value-pairs mapping the unique person identifier to responses.

zeroval: The zero value for the responses in matrix form (see Details).

Details

PsychometricTest stores the person by item response matrix as a coordinate list, allowing for a common data structure for both dense (all persons respond to the same items) and sparse tests (persons respond to (a subset of) different items, e.g. test equating). For sparse tests or tests with missing responses, zeroval determines the value of the missing responses when reconstructing the response matrix via Matrix.

Construction

At construction item and person pointers are precomputed to allow efficient lookup of responses for both items and persons.

Warning

It is required that all item ids in items as well as person ids in person are unique! Otherwise an ArgumentError is thrown.

PsychometricTests.PsychometricTest — Method

PsychometricTest(items, persons, responses)

Construct a psychometric test from a vector of items, a vector of persons, and a vector of responses.

Warning

It is required that all item ids in items as well as person ids in person are unique! Otherwise an ArgumentError is thrown.

Examples

julia> items = [BasicItem(i) for i in 1:3];

julia> persons = [BasicPerson(p) for p in 1:2];

julia> responses = vec([BasicResponse(i, p, rand(0:1)) for i in 1:3, p in 1:2]);

julia> test = PsychometricTest(items, persons, responses)
A PsychometricTest with 6 BasicResponse{Int64, Int64, Int64} from 2 BasicPerson{Int64} and 3 BasicItem{Int64}

PsychometricTests.PsychometricTest — Method

PsychometricTest(data::AbstractMatrix)

Construct a psychometric test from a response matrix.

Id variables for persons and items are derived from the row number and column number respectively.

Examples

julia> response_data = rand(0:1, 5, 3);

julia> test = PsychometricTest(response_data)
A PsychometricTest with 15 BasicResponse{Int64, Int64, Int64} from 5 BasicPerson{Int64} and 3 BasicItem{Int64}

PsychometricTests.PsychometricTest — Method

PsychometricTest(table, item_vars = Tables.columnnames(table), id_var = nothing)

Construct a psychometric test from a Tables.jl compatible table.

If only a column subset of table is to be included in the resulting test, specify a vector of column names in item_vars. The unique item identifier then corresponds to the column name of the table.

id_var is used to determine the unique person identificators. If id_var = nothing (the default), the person id is an integer index corresponding to the table row.

Examples

From DataFrame with default settings

julia> using DataFrames

julia> response_data = DataFrame(Item1 = [1, 0, 1], Item2 = [0, 0, 1]);

julia> test = PsychometricTest(response_data)
A PsychometricTest with 6 BasicResponse{Symbol, Int64, Int64} from 3 BasicPerson{Int64} and 2 BasicItem{Symbol}

From a subset of a DataFrame

julia> using DataFrames

julia> response_data = DataFrame(Item1 = [1, 0, 1], Item2 = [0, 0, 1], Item3 = [1, 1, 1]);

julia> test = PsychometricTest(response_data, [:Item1, :Item3])
A PsychometricTest with 6 BasicResponse{Symbol, Int64, Int64} from 3 BasicPerson{Int64} and 2 BasicItem{Symbol}

From a DataFrame with custom id variable

julia> using DataFrames

julia> response_data = DataFrame(id = [:a, :b, :c], Item1 = [1, 0, 1], Item2 = [0, 0, 1]);

julia> test = PsychometricTest(response_data, [:Item1, :Item2], :id)
A PsychometricTest with 6 BasicResponse{Symbol, Symbol, Int64} from 3 BasicPerson{Symbol} and 2 BasicItem{Symbol}

PsychometricTests.Item — Type

Item

An abstract type representing an item in a psychometric test.

Every implementation T <: Item must define the following interface:

getid: Get the unique item identifier.

PsychometricTests.Person — Type

Person

An abstract type representing a person in a psychometric test.

Every implementation T <: Person must define the following interface:

getid: Get the unique person identifier.

PsychometricTests.Response — Type

Response

An abstract type representing a response in a psychometric test.

Every implementation of T <: Response must define the following interface:

getvalue: Get the response value.

getitemid: Get the item id.

getpersonid: Get the person id.

PsychometricTests.BasicItem — Type

BasicItem{I} <: Item

A minimal implementation of Item. Contains no information besides a unique id of type I.

PsychometricTests.BasicPerson — Type

BasicPerson{I} <: Person

A minimal implementation of Person. Contains no information besides a unique id of type I.

PsychometricTests.BasicResponse — Type

BasicResponse{IIT,PIT,T}

A minimal implementation of Response. Contains an item id item_id::IIT, person id person_id::PIT and a response value value::T.

Base.Matrix — Type

Matrix(test::PsychometricTest)

Get the person by item response matrix from test.

Examples

julia> response_data = [0 1 0; 1 0 0; 0 0 1];

julia> test = PsychometricTest(response_data);

julia> Matrix(test)
3×3 Matrix{Int64}:
 0  1  0
 1  0  0
 0  0  1

julia> Matrix(test) == response_data
true

Accessors

PsychometricTests.getitems — Function

getitems(test::PsychometricTest)

Get the vector of items of a psychometric test.

Examples

julia> data = rand(0:1, 3, 2);

julia> test = PsychometricTest(data);

julia> getitems(test)
2-element Vector{BasicItem{Int64}}:
 BasicItem{Int64}(1)
 BasicItem{Int64}(2)

PsychometricTests.getpersons — Function

getpersons(test::PsychometricTest)

Get the vector of persons of a psychometric test.

Examples

julia> data = rand(0:1, 3, 2);

julia> test = PsychometricTest(data);

julia> getpersons(test)
3-element Vector{BasicPerson{Int64}}:
 BasicPerson{Int64}(1)
 BasicPerson{Int64}(2)
 BasicPerson{Int64}(3)

PsychometricTests.getresponses — Function

getresponses(test::PsychometricTest)

Get the vector of responses of a psychometric test.

Examples

julia> data = [0 0 1; 1 0 1];

julia> test = PsychometricTest(data);

julia> getresponses(test)
6-element Vector{BasicResponse{Int64, Int64, Int64}}:
 BasicResponse{Int64, Int64, Int64}(1, 1, 0)
 BasicResponse{Int64, Int64, Int64}(1, 2, 1)
 BasicResponse{Int64, Int64, Int64}(2, 1, 0)
 BasicResponse{Int64, Int64, Int64}(2, 2, 0)
 BasicResponse{Int64, Int64, Int64}(3, 1, 1)
 BasicResponse{Int64, Int64, Int64}(3, 2, 1)

PsychometricTests.getvalue — Function

getvalue(response::Response)

Get the response value of response.

PsychometricTests.getid — Function

getid(item::Item)

Get the unique item identifier of item.

getid(person::Person)

Get the unique person identifier of person.

PsychometricTests.getitemid — Function

getitemid(response::Response)

Get the unique item id of response.

PsychometricTests.getpersonid — Function

getpersonid(response::Response)

get the unique person id of response.

PsychometricTests.nitems — Function

nitems(test::PsychometricTest)

Get the number of items in test.

PsychometricTests.npersons — Function

npersons(test::PsychometricTest)

Get the number of persons in test.

PsychometricTests.nresponses — Function

nresponses(test::PsychometricTest)

Get the number of responses in test.

Iterators

PsychometricTests.eachitem — Function

eachitem(test::PsychometricTest)

Iterate over each item in test.

PsychometricTests.eachperson — Function

eachperson(test::PsychometricTest)

Iterate over each person in test.

PsychometricTests.eachresponse — Function

eachresponse(test::PsychometricTest)

Iterate over each response in test.

Descriptive Statistics

PsychometricTests.itemmean — Function

itemmean(test::PsychometricTest, id)

Calculate the mean score for item with identifier id in test.

PsychometricTests.itemmeans — Function

itemmeans(test::PsychometricTest)

Calculate the mean scores for each item in test.

PsychometricTests.itemscore — Function

itemscore(test::PsychometricTest, id)

Calculate the total score for item with identifier id in test.

PsychometricTests.itemscores — Function

itemscores(test::PsychometricTest)

Calculate the total score for each item in test.

PsychometricTests.personmean — Function

personmean(test::PsychometricTest, id)

Calculate the mean score for person with identifier id in test.

PsychometricTests.personmeans — Function

personmeans(test::PsychometricTest)

Calculate the mean scores for each person in test.

PsychometricTests.personscore — Function

personscore(test::PsychometricTest, id)

Calculate the total score for person with identifier id in test.

PsychometricTests.personscores — Function

personscores(test::PsychometricTest)

Calculate the total scores for each person in test.